feat: update inventory view
